add doxygen docs and convert MyPage template class from header only to header plus implementation
