+++ doc/doxygen/Doxyfile.in
@@ -410,24 +410,27 @@ WARN_LOGFILE           = "../doxygen-warn.log"
 # spaces. See also FILE_PATTERNS and EXTENSION_MAPPING
 # Note: If this tag is empty the current directory is searched.
 
-INPUT                  = @LAMMPS_SOURCE_DIR@/utils.cpp      \
-                         @LAMMPS_SOURCE_DIR@/utils.h        \
-                         @LAMMPS_SOURCE_DIR@/library.cpp    \
-                         @LAMMPS_SOURCE_DIR@/library.h      \
-                         @LAMMPS_SOURCE_DIR@/lammps.cpp     \
-                         @LAMMPS_SOURCE_DIR@/lammps.h       \
-                         @LAMMPS_SOURCE_DIR@/lmptype.h      \
-                         @LAMMPS_SOURCE_DIR@/pointers.h     \
-                         @LAMMPS_SOURCE_DIR@/atom.cpp       \
-                         @LAMMPS_SOURCE_DIR@/atom.h         \
-                         @LAMMPS_SOURCE_DIR@/input.cpp      \
-                         @LAMMPS_SOURCE_DIR@/input.h        \
-                         @LAMMPS_SOURCE_DIR@/tokenizer.cpp  \
-                         @LAMMPS_SOURCE_DIR@/tokenizer.h    \
-                         @LAMMPS_SOURCE_DIR@/text_file_reader.cpp  \
-                         @LAMMPS_SOURCE_DIR@/text_file_reader.h    \
-                         @LAMMPS_SOURCE_DIR@/potential_file_reader.cpp  \
-                         @LAMMPS_SOURCE_DIR@/potential_file_reader.h    \
+INPUT                  = @LAMMPS_SOURCE_DIR@/utils.cpp                 \
+                         @LAMMPS_SOURCE_DIR@/utils.h                   \
+                         @LAMMPS_SOURCE_DIR@/library.cpp               \
+                         @LAMMPS_SOURCE_DIR@/library.h                 \
+                         @LAMMPS_SOURCE_DIR@/lammps.cpp                \
+                         @LAMMPS_SOURCE_DIR@/lammps.h                  \
+                         @LAMMPS_SOURCE_DIR@/lmptype.h                 \
+                         @LAMMPS_SOURCE_DIR@/pointers.h                \
+                         @LAMMPS_SOURCE_DIR@/atom.cpp                  \
+                         @LAMMPS_SOURCE_DIR@/atom.h                    \
+                         @LAMMPS_SOURCE_DIR@/input.cpp                 \
+                         @LAMMPS_SOURCE_DIR@/input.h                   \
+                         @LAMMPS_SOURCE_DIR@/tokenizer.cpp             \
+                         @LAMMPS_SOURCE_DIR@/tokenizer.h               \
+                         @LAMMPS_SOURCE_DIR@/text_file_reader.cpp      \
+                         @LAMMPS_SOURCE_DIR@/text_file_reader.h        \
+                         @LAMMPS_SOURCE_DIR@/potential_file_reader.cpp \
+                         @LAMMPS_SOURCE_DIR@/potential_file_reader.h   \
+                         @LAMMPS_SOURCE_DIR@/my_page.cpp               \
+                         @LAMMPS_SOURCE_DIR@/my_page.h                 \
+                         @LAMMPS_SOURCE_DIR@/my_pool_chunk.h           \
 
 # The EXCLUDE_SYMLINKS tag can be used to select whether or not files or
 # directories that are symbolic links (a Unix file system feature) are excluded
