+++ src/my_page.h
@@ -12,37 +12,7 @@
 ------------------------------------------------------------------------- */
 
 /* ----------------------------------------------------------------------
-MyPage = templated class for storing chunks of datums in pages
-  chunks are not returnable, can only reset and start over
-  replaces many small mallocs with a few large mallocs
-  pages are never freed, so can reuse w/out reallocs
-usage:
-  request one datum at a time, repeat, clear
-  request chunks of datums in each get() or vget(), repeat, clear
-  chunk size can vary from request to request
-  chunk size can be known in advance or registered after usage via vgot()
-inputs:
-   template T = one datum, e.g. int, double, struct, int[3]
-     for int[3], access datum as ivec[i][2]
-methods:
-   T *get() = return ptr to one datum
-   T *get(N) = return ptr to N datums, N < maxchunk required
-   T *vget() = return ptr to maxchunk datums, use as needed, then call vgot()
-     all gets return NULL if error encountered
-   vgot(N) = used N datums of previous vget(), N < maxchunk required
-   void init(maxchunk, pagesize, pagedelta)
-     define allocation params and allocate first page(s)
-     call right after constructor
-       can call again to reset allocation params and free previous pages
-     maxchunk = max # of datums in one chunk, default = 1
-     pagesize = # of datums in one page, default = 1024
-       should be big enough to store multiple chunks
-     pagedelta = # of pages to allocate at a time, default = 1
-     return 1 if bad params
-   void reset() = clear pages w/out freeing
-   int size() = return total size of allocated pages in bytes
-   int status() = return error status
-     0 = ok, 1 = chunksize > maxchunk, 2 = allocation error
+   templated class for storing chunks of datums in pages
 ------------------------------------------------------------------------- */
 
 #ifndef LAMMPS_MY_PAGE_H
@@ -52,7 +22,6 @@ methods:
 #define LAMMPS_MEMALIGN 64
 #endif
 
-#include <cstdlib>
 namespace LAMMPS_NS {
 
 template<class T>
@@ -60,139 +29,33 @@ class MyPage {
  public:
   int ndatum;      // total # of stored datums
   int nchunk;      // total # of stored chunks
-
-  MyPage() {
-    ndatum = nchunk = 0;
-    pages = NULL;
-    npage = 0;
-    errorflag = 0;
-  }
-
-  // (re)initialize allocation params
-  // also allocate first page(s)
+  MyPage();
+  virtual ~MyPage();
 
   int init(int user_maxchunk = 1, int user_pagesize = 1024,
-           int user_pagedelta = 1) {
-    maxchunk = user_maxchunk;
-    pagesize = user_pagesize;
-    pagedelta = user_pagedelta;
+           int user_pagedelta = 1);
+  
+  T *get();
+  T *get(int n);
 
-    if (maxchunk <= 0 || pagesize <= 0 || pagedelta <= 0) return 1;
-    if (maxchunk > pagesize) return 1;
+  T *vget();
+  void vgot(int n);
 
-    // free any previously allocated pages
+  void reset();
 
-    for (int i = 0; i < npage; i++) free(pages[i]);
-    free(pages);
-
-    // initial page allocation
-
-    ndatum = nchunk = 0;
-    pages = NULL;
-    npage = 0;
-    allocate();
-    if (errorflag) return 2;
-    ipage = index = 0;
-    page = pages[ipage];
-    return 0;
-  }
-
-  // free all allocated pages
-
-  ~MyPage() {
-    for (int i = 0; i < npage; i++) free(pages[i]);
-    free(pages);
-  }
-
-  // get ptr to one datum
-  // return NULL if run out of memory
-
-  T *get() {
-    ndatum++;
-    nchunk++;
-    if (index < pagesize) return &page[index++];
-    ipage++;
-    if (ipage == npage) {
-      allocate();
-      if (errorflag) return NULL;
-    }
-    page = pages[ipage];
-    index = 0;
-    return &page[index++];
-  }
-
-  // get ptr to location that can store N datums
-  // error if N > maxchunk
-  // return NULL if run out of memory
-
-  T *get(int n) {
-    if (n > maxchunk) {
-      errorflag = 1;
-      return NULL;
-    }
-    ndatum += n;
-    nchunk++;
-    if (index+n <= pagesize) {
-      int start = index;
-      index += n;
-      return &page[start];
-    }
-    ipage++;
-    if (ipage == npage) {
-      allocate();
-      if (errorflag) return NULL;
-    }
-    page = pages[ipage];
-    index = n;
-    return &page[0];
-  }
-
-  // get ptr to location that can store maxchunk datums
-  // will return same ptr as previous call if vgot() not called
-  // return NULL if run out of memory
-
-  T *vget() {
-    if (index+maxchunk <= pagesize) return &page[index];
-    ipage++;
-    if (ipage == npage) {
-      allocate();
-      if (errorflag) return NULL;
-    }
-    page = pages[ipage];
-    index = 0;
-    return &page[index];
-  }
-
-  // increment by N = # of values stored in loc returned by vget()
-  // OK to not call if vget() ptr was not used
-  // error if N > maxchunk
-
-  void vgot(int n) {
-    if (n > maxchunk) errorflag = 1;
-    ndatum += n;
-    nchunk++;
-    index += n;
-  }
-
-  // clear all pages, without freeing any memory
-
-  void reset() {
-    ndatum = nchunk = 0;
-    index = ipage = 0;
-    page = pages[ipage];
-  }
-
-  // return total size of allocated pages
+  /** Return total size of allocated pages
+   *
+   * \return total storage used in bytes */
 
   int size() const {
     return npage*pagesize*sizeof(T);
   }
 
-  // return error status
+  /** Return error status 
+   *
+   * \return 0 if no error, 1 requested chunk size > maxchunk, 2 if malloc failed */
 
-  int status() const {
-    return errorflag;
-  }
+  int status() const { return errorflag; }
 
  private:
   T **pages;      // list of allocated pages
@@ -208,27 +71,7 @@ class MyPage {
   int errorflag;  // flag > 0 if error has occurred
                   // 1 = chunk size exceeded maxchunk
                   // 2 = memory allocation error
-
-  void allocate() {
-    npage += pagedelta;
-    pages = (T **) realloc(pages,npage*sizeof(T *));
-    if (!pages) {
-      errorflag = 2;
-      return;
-    }
-
-    for (int i = npage-pagedelta; i < npage; i++) {
-#if defined(LAMMPS_MEMALIGN)
-      void *ptr;
-      if (posix_memalign(&ptr, LAMMPS_MEMALIGN, pagesize*sizeof(T)))
-        errorflag = 2;
-      pages[i] = (T *) ptr;
-#else
-      pages[i] = (T *) malloc(pagesize*sizeof(T));
-      if (!pages[i]) errorflag = 2;
-#endif
-    }
-  }
+  void allocate();
 };
 
 }
